test: move lock/unlock wallet tests from Wallet.test to Wallets.test
