+++ src/components/Wallet.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { render, screen, waitFor, waitForElementToBeRemoved } from '../testUtils'
+import { render, screen, waitFor } from '../testUtils'
 import { act } from 'react-dom/test-utils'
 import user from '@testing-library/user-event'
 import * as apiMock from '../libs/JmWalletApi'
@@ -10,45 +10,31 @@ import Wallet from './Wallet'
 jest.mock('../libs/JmWalletApi', () => ({
   ...jest.requireActual('../libs/JmWalletApi'),
   getSession: jest.fn(),
-  postWalletUnlock: jest.fn(),
-  getWalletLock: jest.fn(),
 }))
-
-const mockedNavigate = jest.fn()
-jest.mock('react-router-dom', () => {
-  return {
-    ...jest.requireActual('react-router-dom'),
-    useNavigate: () => mockedNavigate,
-  }
-})
 
 describe('<Wallet />', () => {
   const dummyWalletName = 'dummy.jmdat'
-  const dummyToken = 'dummyToken'
+  const dummyPassword = 'correct horse battery staple'
 
-  const mockStartWallet = jest.fn()
-  const mockStopWallet = jest.fn()
-  const mockSetAlert = jest.fn()
+  const mockUnlockWallet = jest.fn()
+  const mockLockWallet = jest.fn()
 
   const setup = ({
     name,
+    lockWallet = undefined,
+    unlockWallet = undefined,
     isActive = false,
-    hasToken = false,
     makerRunning = false,
     coinjoinInProgress = false,
-    currentWallet = null,
   }) => {
     render(
       <Wallet
         name={name}
+        lockWallet={lockWallet}
+        unlockWallet={unlockWallet}
         isActive={isActive}
-        hasToken={hasToken}
-        makerRunning={makerRunning}
         coinjoinInProgress={coinjoinInProgress}
-        currentWallet={currentWallet}
-        startWallet={mockStartWallet}
-        stopWallet={mockStopWallet}
-        setAlert={mockSetAlert}
+        makerRunning={makerRunning}
       />
     )
   }
@@ -63,29 +49,23 @@ describe('<Wallet />', () => {
 
     expect(screen.getByText(walletDisplayName(dummyWalletName))).toBeInTheDocument()
     expect(screen.getByText('wallets.wallet_preview.wallet_inactive')).toBeInTheDocument()
-    expect(screen.getByPlaceholderText('wallets.wallet_preview.placeholder_password')).toBeInTheDocument()
-    expect(screen.getByText('wallets.wallet_preview.button_unlock')).toBeInTheDocument()
+    expect(screen.queryByPlaceholderText('wallets.wallet_preview.placeholder_password')).not.toBeInTheDocument()
+    expect(screen.queryByText('wallets.wallet_preview.button_unlock')).not.toBeInTheDocument()
     expect(screen.queryByText('wallets.wallet_preview.button_open')).not.toBeInTheDocument()
     expect(screen.queryByText('wallets.wallet_preview.button_lock')).not.toBeInTheDocument()
   })
 
   it('should unlock inactive wallet successfully', async () => {
-    apiMock.postWalletUnlock.mockResolvedValueOnce({
-      ok: true,
-      json: () => Promise.resolve({ walletname: dummyWalletName, token: dummyToken }),
-    })
-
-    await act(async () => setup({ name: dummyWalletName }))
+    await act(async () => setup({ name: dummyWalletName, unlockWallet: mockUnlockWallet }))
 
     expect(screen.getByText('wallets.wallet_preview.wallet_inactive')).toBeInTheDocument()
     expect(screen.getByText('wallets.wallet_preview.button_unlock')).toBeInTheDocument()
     expect(screen.getByPlaceholderText('wallets.wallet_preview.placeholder_password')).toBeInTheDocument()
+    expect(screen.queryByText('wallets.wallet_preview.button_open')).not.toBeInTheDocument()
+    expect(screen.queryByText('wallets.wallet_preview.button_lock')).not.toBeInTheDocument()
 
     act(() => {
-      user.type(
-        screen.getByPlaceholderText('wallets.wallet_preview.placeholder_password'),
-        'correct horse battery staple'
-      )
+      user.paste(screen.getByPlaceholderText('wallets.wallet_preview.placeholder_password'), dummyPassword)
     })
 
     await act(async () => {
@@ -96,49 +76,34 @@ describe('<Wallet />', () => {
       await waitFor(() => screen.findByText('wallets.wallet_preview.button_unlock'))
     })
 
-    expect(mockStartWallet).toHaveBeenCalledWith(dummyWalletName, dummyToken)
-    expect(mockedNavigate).toHaveBeenCalledWith('/wallet')
+    expect(mockUnlockWallet).toHaveBeenCalledWith(dummyWalletName, dummyPassword)
   })
 
-  it('should add alert if unlocking of inactive wallet fails', async () => {
-    const apiErrorMessage = 'ANY_ERROR_MESSAGE with template <Wallet>'
-    apiMock.postWalletUnlock.mockResolvedValueOnce({
-      ok: false,
-      json: () => Promise.resolve({ message: apiErrorMessage }),
-    })
+  it('should provide ability to unlock active wallet (i.e. when auth token is missing)', () => {
+    act(() =>
+      setup({
+        name: dummyWalletName,
+        isActive: true,
+        unlockWallet: mockUnlockWallet,
+      })
+    )
 
-    act(() => setup({ name: dummyWalletName }))
-
-    expect(screen.getByText('wallets.wallet_preview.wallet_inactive')).toBeInTheDocument()
-    expect(screen.getByText('wallets.wallet_preview.button_unlock')).toBeInTheDocument()
-    expect(screen.getByPlaceholderText('wallets.wallet_preview.placeholder_password')).toBeInTheDocument()
-
-    act(() => {
-      user.type(
-        screen.getByPlaceholderText('wallets.wallet_preview.placeholder_password'),
-        'correct horse battery staple'
-      )
-    })
-
-    await act(async () => {
-      const unlockWalletButton = screen.getByText('wallets.wallet_preview.button_unlock')
-      user.click(unlockWalletButton)
-
-      await waitFor(() => screen.findByText(/wallets.wallet_preview.button_unlocking/))
-      await waitFor(() => screen.findByText('wallets.wallet_preview.button_unlock'))
-    })
-
-    expect(mockStartWallet).not.toHaveBeenCalled()
-    expect(mockedNavigate).not.toHaveBeenCalled()
-    expect(mockSetAlert).toHaveBeenCalledWith({
-      variant: 'danger',
-      dismissible: false,
-      message: apiErrorMessage.replace('Wallet', dummyWalletName),
-    })
+    expect(screen.getByText(walletDisplayName(dummyWalletName))).toBeInTheDocument()
+    expect(screen.getByText('wallets.wallet_preview.wallet_active')).toBeInTheDocument()
+    expect(screen.queryByPlaceholderText('wallets.wallet_preview.placeholder_password')).toBeInTheDocument()
+    expect(screen.queryByText('wallets.wallet_preview.button_unlock')).toBeInTheDocument()
+    expect(screen.queryByText('wallets.wallet_preview.button_open')).not.toBeInTheDocument()
+    expect(screen.queryByText('wallets.wallet_preview.button_lock')).not.toBeInTheDocument()
   })
 
   it('should render active wallet without errors', () => {
-    act(() => setup({ name: dummyWalletName, isActive: true, hasToken: true }))
+    act(() =>
+      setup({
+        name: dummyWalletName,
+        isActive: true,
+        lockWallet: mockLockWallet,
+      })
+    )
 
     expect(screen.getByText(walletDisplayName(dummyWalletName))).toBeInTheDocument()
     expect(screen.getByText('wallets.wallet_preview.wallet_active')).toBeInTheDocument()
@@ -149,17 +114,11 @@ describe('<Wallet />', () => {
   })
 
   it('should lock active wallet successfully', async () => {
-    apiMock.getWalletLock.mockResolvedValueOnce({
-      ok: true,
-      json: () => Promise.resolve({ walletname: dummyWalletName, already_locked: false }),
-    })
-
     act(() =>
       setup({
         name: dummyWalletName,
         isActive: true,
-        hasToken: true,
-        currentWallet: { name: dummyWalletName, token: dummyToken },
+        lockWallet: mockLockWallet,
       })
     )
 
@@ -174,97 +133,7 @@ describe('<Wallet />', () => {
       await waitFor(() => screen.findByText('wallets.wallet_preview.button_lock'))
     })
 
-    expect(mockStopWallet).toHaveBeenCalled()
-    expect(mockSetAlert).toHaveBeenCalledWith({
-      variant: 'success',
-      dismissible: false,
-      message: `wallets.wallet_preview.alert_wallet_locked_successfully`,
-    })
-  })
-
-  it('should add alert if locking active wallet fails', async () => {
-    const apiErrorMessage = 'ANY_ERROR_MESSAGE'
-    apiMock.getWalletLock.mockResolvedValueOnce({
-      ok: false,
-      status: 500, // any other error status than 401
-      json: () => Promise.resolve({ message: apiErrorMessage }),
-    })
-
-    act(() =>
-      setup({
-        name: dummyWalletName,
-        isActive: true,
-        hasToken: true,
-        currentWallet: { name: dummyWalletName, token: dummyToken },
-      })
-    )
-
-    expect(screen.getByText('wallets.wallet_preview.wallet_active')).toBeInTheDocument()
-    expect(screen.getByText('wallets.wallet_preview.button_lock')).toBeInTheDocument()
-
-    await act(async () => {
-      const lockWalletButton = screen.getByText('wallets.wallet_preview.button_lock')
-      user.click(lockWalletButton)
-
-      await waitFor(() => screen.findByText(/wallets.wallet_preview.button_locking/))
-      await waitFor(() => screen.findByText('wallets.wallet_preview.button_lock'))
-    })
-
-    // on errors other than 401, stopWallet should not have been called
-    expect(mockStopWallet).not.toHaveBeenCalled()
-    expect(mockSetAlert).toHaveBeenCalledWith({
-      variant: 'danger',
-      dismissible: false,
-      message: apiErrorMessage,
-    })
-  })
-
-  it('should add alert but clear wallet if locking active wallet fails with UNAUTHORIZED', async () => {
-    const apiErrorMessage = 'Invalid credentials.'
-    apiMock.getWalletLock.mockResolvedValueOnce({
-      ok: false,
-      status: 401,
-      json: () => Promise.resolve({ message: apiErrorMessage }),
-    })
-
-    act(() =>
-      setup({
-        name: dummyWalletName,
-        isActive: true,
-        hasToken: true,
-        currentWallet: { name: dummyWalletName, token: dummyToken },
-      })
-    )
-
-    expect(screen.getByText('wallets.wallet_preview.wallet_active')).toBeInTheDocument()
-    expect(screen.getByText('wallets.wallet_preview.button_lock')).toBeInTheDocument()
-
-    await act(async () => {
-      const lockWalletButton = screen.getByText('wallets.wallet_preview.button_lock')
-      user.click(lockWalletButton)
-
-      await waitFor(() => screen.findByText(/wallets.wallet_preview.button_locking/))
-      await waitFor(() => screen.findByText('wallets.wallet_preview.button_lock'))
-    })
-
-    // on 401 errors, stopWallet *should* have been called
-    expect(mockStopWallet).toHaveBeenCalled()
-    expect(mockSetAlert).toHaveBeenCalledWith({
-      variant: 'danger',
-      dismissible: false,
-      message: apiErrorMessage,
-    })
-  })
-
-  it('should provide ability to unlock active wallet when token is missing', () => {
-    act(() => setup({ name: dummyWalletName, isActive: true, hasToken: false }))
-
-    expect(screen.getByText(walletDisplayName(dummyWalletName))).toBeInTheDocument()
-    expect(screen.getByText('wallets.wallet_preview.wallet_active')).toBeInTheDocument()
-    expect(screen.queryByPlaceholderText('wallets.wallet_preview.placeholder_password')).toBeInTheDocument()
-    expect(screen.queryByText('wallets.wallet_preview.button_unlock')).toBeInTheDocument()
-    expect(screen.queryByText('wallets.wallet_preview.button_open')).not.toBeInTheDocument()
-    expect(screen.queryByText('wallets.wallet_preview.button_lock')).not.toBeInTheDocument()
+    expect(mockLockWallet).toHaveBeenCalledWith(dummyWalletName, { confirmed: false })
   })
 
   it('should display earn indicator when maker is running', () => {
@@ -280,71 +149,4 @@ describe('<Wallet />', () => {
     expect(document.querySelector('.joining-indicator')).toBeInTheDocument()
     expect(document.querySelector('.earn-indicator')).toBeNull()
   })
-
-  it.each`
-    makerRunning | coinjoinInProgress | expectedModalBody
-    ${true}      | ${false}           | ${'wallets.wallet_preview.modal_lock_wallet_maker_running_text wallets.wallet_preview.modal_lock_wallet_alternative_action_text'}
-    ${false}     | ${true}            | ${'wallets.wallet_preview.modal_lock_wallet_coinjoin_in_progress_text wallets.wallet_preview.modal_lock_wallet_alternative_action_text'}
-  `(
-    'should confirm locking wallet if maker ($makerRunning) or taker ($coinjoinInProgress) is running',
-    async ({ makerRunning, coinjoinInProgress, expectedModalBody }) => {
-      apiMock.getWalletLock.mockResolvedValueOnce({
-        ok: true,
-        json: () => Promise.resolve({ walletname: dummyWalletName, already_locked: false }),
-      })
-
-      act(() =>
-        setup({
-          name: dummyWalletName,
-          isActive: true,
-          hasToken: true,
-          makerRunning,
-          coinjoinInProgress,
-          currentWallet: { name: dummyWalletName, token: dummyToken },
-        })
-      )
-
-      // modal is initially not shown
-      expect(screen.queryByText('wallets.wallet_preview.modal_lock_wallet_title')).not.toBeInTheDocument()
-      expect(screen.getByText('wallets.wallet_preview.button_lock')).toBeInTheDocument()
-
-      await act(async () => {
-        // click on the "lock" button
-        const lockWalletButton = screen.getByText('wallets.wallet_preview.button_lock')
-        user.click(lockWalletButton)
-      })
-
-      // modal appeared
-      expect(screen.getByText('wallets.wallet_preview.modal_lock_wallet_title')).toBeInTheDocument()
-      expect(screen.getByText(expectedModalBody)).toBeInTheDocument()
-      expect(screen.getByText('modal.confirm_button_accept')).toBeInTheDocument()
-      expect(screen.getByText('modal.confirm_button_reject')).toBeInTheDocument()
-
-      act(() => {
-        // click on the modal's "cancel" button
-        const lockWalletButton = screen.getByText('modal.confirm_button_reject')
-        user.click(lockWalletButton)
-      })
-      await waitForElementToBeRemoved(screen.getByText('wallets.wallet_preview.modal_lock_wallet_title'))
-
-      expect(mockStopWallet).not.toHaveBeenCalled()
-
-      await act(async () => {
-        const lockWalletButton = screen.getByText('wallets.wallet_preview.button_lock')
-        user.click(lockWalletButton)
-      })
-
-      // modal appeared
-      expect(screen.getByText('wallets.wallet_preview.modal_lock_wallet_title')).toBeInTheDocument()
-
-      act(() => {
-        // click on the modal's "confirm" button
-        const lockWalletButton = screen.getByText('modal.confirm_button_accept')
-        user.click(lockWalletButton)
-      })
-      await waitForElementToBeRemoved(screen.getByText('wallets.wallet_preview.modal_lock_wallet_title'))
-
-      expect(mockStopWallet).toHaveBeenCalled()
-    }
-  )
 })
